replaced doubles with decimals

+++ LeagueOptimizer.Models/Champions/Caitlyn/AbilityData/CaitlynAbilityData.cs
@@ -12,8 +12,8 @@ public class CaitlynAbilityData
     {
         public List<int> ChampionBaseDmg { get; set; }
         public List<int> BaseDmg { get; set; }
-        public double CritScaling { get; set; }
-        public double IeCritScaling { get; set; }
+        public decimal CritScaling { get; set; }
+        public decimal IeCritScaling { get; set; }
     }
 
     public class SpellQData
@@ -21,19 +21,19 @@ public class CaitlynAbilityData
         public List<int> Cost { get; set; }
         public List<int> Cooldown { get; set; }
         public List<int> BaseDmg { get; set; }
-        public List<double> TotalAdScaling { get; set; }
-        public double ReducedDamageMultiplier { get; set; }
+        public List<decimal> TotalAdScaling { get; set; }
+        public decimal ReducedDamageMultiplier { get; set; }
     }
 
     public class SpellWData
     {
         public int Cost { get; set; }
-        public double Cooldown { get; set; }
+        public decimal Cooldown { get; set; }
         public List<int> Recharge { get; set; }
         public List<int> Duration { get; set; }
         public List<int> MaximumCharges { get; set; }
         public List<int> BaseDmg { get; set; }
-        public double BonusAdScaling { get; set; }
+        public decimal BonusAdScaling { get; set; }
     }
 
     public class SpellEData
@@ -41,9 +41,9 @@ public class CaitlynAbilityData
         public int Cost { get; set; }
         public List<int> Cooldown { get; set; }
         public List<int> BaseDmg { get; set; }
-        public double ApScaling { get; set; }
-        public double SlowAmount { get; set; }
-        public double SlowDuration { get; set; }
+        public decimal ApScaling { get; set; }
+        public decimal SlowAmount { get; set; }
+        public decimal SlowDuration { get; set; }
     }
 
     public class SpellRData
@@ -51,7 +51,7 @@ public class CaitlynAbilityData
         public int Cost { get; set; }
         public int Cooldown { get; set; }
         public List<int> BaseDmg { get; set; }
-        public double BonusAdScaling { get; set; }
-        public double CritScaling { get; set; }
+        public decimal BonusAdScaling { get; set; }
+        public decimal CritScaling { get; set; }
     }
 }
